v1/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
google-cloud-bigquery
pyarrow
scipy
